+++ changelog.md
@@ -5,6 +5,7 @@
 ### 2.2.4, soon...
 
 * fixed: logging should not record encrypted card details (even though they're encrypted)
+* fixed: stored payments legacy API fails
 * changed: replace old $field array access with GF_Field object access
 
 ### 2.2.3, 2016-12-08